isQuestion middleware + timestamps
