+++ functions/routes/doubt-route.js
@@ -1,10 +1,11 @@
 const express = require("express");
 const { addDoubt} = require("../controllers/doubtController");
+const {isQuestion} = require('../middleware/isQuestion')
 
 
 const router = express.Router();
 
-router.post('/', addDoubt);
+router.post('/',isQuestion, addDoubt);
 
 module.exports = {
     routes : router
